pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-project-template-backend"
version = "0.1.0"
description = "AI Auto Hub service backend template"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "aiosqlite>=0.20.0,<1.0.0",
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "pydantic-settings>=2.8.0,<3.0.0",
  "sqlalchemy>=2.0.38,<3.0.0",
  "uvicorn[standard]>=0.34.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
  "ruff>=0.11.5,<1.0.0",
  "pytest>=8.3.0,<9.0.0",
  "pytest-asyncio>=0.25.0,<1.0.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
